Refactor utils
