+++ android/app/src/main/java/com/microsoft/codepush/react/managers/CodePushTelemetryManagerDeserializer.java
@@ -2,8 +2,7 @@ package com.microsoft.codepush.react.managers;
 
 import com.facebook.react.bridge.WritableMap;
 import com.microsoft.codepush.react.datacontracts.CodePushStatusReport;
-import com.microsoft.codepush.react.managers.CodePushTelemetryManager;
-import com.microsoft.codepush.react.utils.CodePushUtils;
+import com.microsoft.codepush.react.utils.CodePushRNUtils;
 
 public class CodePushTelemetryManagerDeserializer {
     private CodePushTelemetryManager mTelemetryManager;
@@ -15,7 +14,7 @@ public class CodePushTelemetryManagerDeserializer {
     public CodePushStatusReport getRollbackReport(WritableMap lastFailedPackage) {
         WritableMap failedStatusReport = mTelemetryManager.getRollbackReport(lastFailedPackage);
         if (failedStatusReport != null) {
-            return CodePushUtils.convertWritableMapToObject(failedStatusReport, CodePushStatusReport.class);
+            return CodePushRNUtils.convertWritableMapToObject(failedStatusReport, CodePushStatusReport.class);
         }
         return null;
     }
@@ -23,7 +22,7 @@ public class CodePushTelemetryManagerDeserializer {
     public CodePushStatusReport getUpdateReport(WritableMap currentPackage) {
         WritableMap newPackageStatusReport = mTelemetryManager.getUpdateReport(currentPackage);
         if (newPackageStatusReport != null) {
-            return CodePushUtils.convertWritableMapToObject(newPackageStatusReport, CodePushStatusReport.class);
+            return CodePushRNUtils.convertWritableMapToObject(newPackageStatusReport, CodePushStatusReport.class);
         }
         return null;
     }
@@ -31,7 +30,7 @@ public class CodePushTelemetryManagerDeserializer {
     public CodePushStatusReport getBinaryUpdateReport(String appVersion) {
         WritableMap newAppVersionStatusReport = mTelemetryManager.getBinaryUpdateReport(appVersion);
         if (newAppVersionStatusReport != null) {
-            return CodePushUtils.convertWritableMapToObject(newAppVersionStatusReport, CodePushStatusReport.class);
+            return CodePushRNUtils.convertWritableMapToObject(newAppVersionStatusReport, CodePushStatusReport.class);
         }
         return null;
     }
@@ -39,7 +38,7 @@ public class CodePushTelemetryManagerDeserializer {
     public CodePushStatusReport getRetryStatusReport() {
         WritableMap retryStatusReport = mTelemetryManager.getRetryStatusReport();
         if (retryStatusReport != null) {
-            return CodePushUtils.convertWritableMapToObject(retryStatusReport, CodePushStatusReport.class);
+            return CodePushRNUtils.convertWritableMapToObject(retryStatusReport, CodePushStatusReport.class);
         }
         return null;
     }
