+++ android/app/src/main/java/com/microsoft/codepush/react/utils/CodePushUtils.java
@@ -2,14 +2,7 @@ package com.microsoft.codepush.react.utils;
 
 import android.util.Log;
 
-import com.facebook.react.bridge.Arguments;
-import com.facebook.react.bridge.NoSuchKeyException;
-import com.facebook.react.bridge.ReadableArray;
-import com.facebook.react.bridge.ReadableMap;
-import com.facebook.react.bridge.ReadableMapKeySetIterator;
-import com.facebook.react.bridge.ReadableType;
-import com.facebook.react.bridge.WritableArray;
-import com.facebook.react.bridge.WritableMap;
+
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonObject;
@@ -33,158 +26,8 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 
-public class CodePushUtils {
+public abstract class CodePushUtils {
     private static Gson mGson = new GsonBuilder().create();
-
-    public static String appendPathComponent(String basePath, String appendPathComponent) {
-        return new File(basePath, appendPathComponent).getAbsolutePath();
-    }
-
-    public static WritableArray convertJsonArrayToWritable(JSONArray jsonArr) {
-        WritableArray arr = Arguments.createArray();
-        for (int i=0; i<jsonArr.length(); i++) {
-            Object obj = null;
-            try {
-                obj = jsonArr.get(i);
-            } catch (JSONException jsonException) {
-                // Should not happen.
-                throw new CodePushUnknownException(i + " should be within bounds of array " + jsonArr.toString(), jsonException);
-            }
-
-            if (obj instanceof JSONObject)
-                arr.pushMap(convertJsonObjectToWritable((JSONObject) obj));
-            else if (obj instanceof JSONArray)
-                arr.pushArray(convertJsonArrayToWritable((JSONArray) obj));
-            else if (obj instanceof String)
-                arr.pushString((String) obj);
-            else if (obj instanceof Double)
-                arr.pushDouble((Double) obj);
-            else if (obj instanceof Integer)
-                arr.pushInt((Integer) obj);
-            else if (obj instanceof Boolean)
-                arr.pushBoolean((Boolean) obj);
-            else if (obj == null)
-                arr.pushNull();
-            else
-                throw new CodePushUnknownException("Unrecognized object: " + obj);
-        }
-
-        return arr;
-    }
-
-    public static WritableMap convertJsonObjectToWritable(JSONObject jsonObj) {
-        WritableMap map = Arguments.createMap();
-        Iterator<String> it = jsonObj.keys();
-        while(it.hasNext()){
-            String key = it.next();
-            Object obj = null;
-            try {
-                obj = jsonObj.get(key);
-            } catch (JSONException jsonException) {
-                // Should not happen.
-                throw new CodePushUnknownException("Key " + key + " should exist in " + jsonObj.toString() + ".", jsonException);
-            }
-
-            if (obj instanceof JSONObject)
-                map.putMap(key, convertJsonObjectToWritable((JSONObject) obj));
-            else if (obj instanceof JSONArray)
-                map.putArray(key, convertJsonArrayToWritable((JSONArray) obj));
-            else if (obj instanceof String)
-                map.putString(key, (String) obj);
-            else if (obj instanceof Double)
-                map.putDouble(key, (Double) obj);
-            else if (obj instanceof Integer)
-                map.putInt(key, (Integer) obj);
-            else if (obj instanceof Boolean)
-                map.putBoolean(key, (Boolean) obj);
-            else if (obj == null)
-                map.putNull(key);
-            else
-                throw new CodePushUnknownException("Unrecognized object: " + obj);
-        }
-
-        return map;
-    }
-
-    public static JSONArray convertReadableToJsonArray(ReadableArray arr) {
-        JSONArray jsonArr = new JSONArray();
-        for (int i=0; i<arr.size(); i++) {
-            ReadableType type = arr.getType(i);
-            switch (type) {
-                case Map:
-                    jsonArr.put(convertReadableToJsonObject(arr.getMap(i)));
-                    break;
-                case Array:
-                    jsonArr.put(convertReadableToJsonArray(arr.getArray(i)));
-                    break;
-                case String:
-                    jsonArr.put(arr.getString(i));
-                    break;
-                case Number:
-                    Double number = arr.getDouble(i);
-                    if ((number == Math.floor(number)) && !Double.isInfinite(number)) {
-                        // This is a whole number.
-                        jsonArr.put(number.longValue());
-                    } else {
-                        try {
-                            jsonArr.put(number.doubleValue());
-                        } catch (JSONException jsonException) {
-                            throw new CodePushUnknownException("Unable to put value " + arr.getDouble(i) + " in JSONArray");
-                        }
-                    }
-                    break;
-                case Boolean:
-                    jsonArr.put(arr.getBoolean(i));
-                    break;
-                case Null:
-                    jsonArr.put(null);
-                    break;
-            }
-        }
-
-        return jsonArr;
-    }
-
-    public static <T> T convertReadableToObject(ReadableMap map,  Class<T> classOfT) {
-        return convertJsonObjectToObject(convertReadableToJsonObject(map), classOfT);
-    }
-
-    public static JSONObject convertReadableToJsonObject(ReadableMap map) {
-        JSONObject jsonObj = new JSONObject();
-        ReadableMapKeySetIterator it = map.keySetIterator();
-        while (it.hasNextKey()) {
-            String key = it.nextKey();
-            ReadableType type = map.getType(key);
-            try {
-                switch (type) {
-                    case Map:
-                        jsonObj.put(key, convertReadableToJsonObject(map.getMap(key)));
-                        break;
-                    case Array:
-                        jsonObj.put(key, convertReadableToJsonArray(map.getArray(key)));
-                        break;
-                    case String:
-                        jsonObj.put(key, map.getString(key));
-                        break;
-                    case Number:
-                        jsonObj.put(key, map.getDouble(key));
-                        break;
-                    case Boolean:
-                        jsonObj.put(key, map.getBoolean(key));
-                        break;
-                    case Null:
-                        jsonObj.put(key, null);
-                        break;
-                    default:
-                        throw new CodePushUnknownException("Unrecognized type: " + type + " of key: " + key);
-                }
-            } catch (JSONException jsonException) {
-                throw new CodePushUnknownException("Error setting key: " + key + " in JSONObject", jsonException);
-            }
-        }
-
-        return jsonObj;
-    }
 
     public static String getStringFromInputStream(InputStream inputStream) throws IOException {
         BufferedReader bufferedReader = null;
@@ -215,27 +58,11 @@ public class CodePushUtils {
         }
     }
 
-    public static void log(String message) {
-        Log.d(CodePushConstants.REACT_NATIVE_LOG_TAG, "[CodePush] " + message);
-    }
-
-    public static void logBundleUrl(String path) {
-        log("Loading JS bundle from \"" + path + "\"");
-    }
-
     public static void setJSONValueForKey(JSONObject json, String key, Object value) {
         try {
             json.put(key, value);
         } catch (JSONException e) {
             throw new CodePushUnknownException("Unable to set value " + value + " for key " + key + " to JSONObject");
-        }
-    }
-
-    public static String tryGetString(ReadableMap map, String key) {
-        try {
-            return map.getString(key);
-        } catch (NoSuchKeyException e) {
-            return null;
         }
     }
 
@@ -261,16 +88,6 @@ public class CodePushUtils {
         return convertStringToObject(jsonObject.toString(), classOfT);
     }
 
-    public static <T> T convertWritableMapToObject(WritableMap writableMap,  Class<T> classOfT) {
-        try {
-            JSONObject jsonObject = new JSONObject(writableMap.toString()).optJSONObject("NativeMap");
-            return convertJsonObjectToObject(jsonObject, classOfT);
-        } catch (JSONException e) {
-            e.printStackTrace();
-            throw new CodePushMalformedDataException(e.toString(), e);
-        }
-    }
-
     public static <T> T convertStringToObject(String stringObject,  Class<T> classOfT) {
         return mGson.fromJson(stringObject, classOfT);
     }
@@ -294,7 +111,4 @@ public class CodePushUtils {
         return sb.toString();
     }
 
-    public static WritableMap convertObjectToWritableMap(Object object) {
-        return convertJsonObjectToWritable(convertObjectToJsonObject(object));
-    }
 }
